back-end/pyproject.toml
[project]
name = "online-therapy-backend"
version = "1.0.0"
description = "Online Therapy Platform - Django REST API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.3.1",
    "djoser>=2.3.3",
    "django-cors-headers>=4.3.1",
    "drf-spectacular>=0.27.2",
    "dj-database-url>=2.1.0",
    "python-dotenv>=1.0.1",
    "django-extensions>=4.1",
    "google-auth>=2.28.0",
    "google-auth-oauthlib>=1.2.0",
    "requests>=2.31.0",
    "Pillow>=10.2.0",
    "cryptography>=42.0.0",
    "sentry-sdk[django]>=1.45.0",
    "django-anymail>=10.3",
    "psycopg2-binary>=2.9.9",
]

[project.optional-dependencies]
dev = [
    "factory-boy>=3.3.0",
    "coverage>=7.4.0",
    "pytest>=8.0.0",
    "pytest-django>=4.8.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["core", "core.settings", "account", "therapy"]
